refactor(ai): replace Semantic Kernel with Microsoft.Extensions.AI
Rewrite the chat implementation behind the unchanged IChatService against
MEAI's IChatClient. SemanticKernelChatService -> ChatService; static kernel
cache -> static IChatClient cache with disposal on eviction. All clients share
the pooled HTTP handler via IHttpClientFactory. Azure api-version is no longer
pinned (the 2.x SDK floats the service version) - AzureApiVersion removed
everywhere.

Carries two type changes made in the same files (neither a logic change):
- Temperature double -> float end-to-end; default 0.2 -> 1.0 (1.0 is the only
  value safe across reasoning models and Claude extended thinking).
- Token counts int? -> long? to match UsageDetails (implicit widening).

Cache metrics tag and gauge renamed kernel -> chat_client. Manual ApiMetrics
instrumentation retained (MEAI OpenTelemetry middleware loses feature
attribution).

+++ TelegramGroupsAdmin.AI/Services/ChatCompletionOptions.cs
@@ -13,7 +13,7 @@ public record ChatCompletionOptions
     /// <summary>
     /// Temperature for randomness (0.0 = deterministic, 2.0 = very random)
     /// </summary>
-    public double? Temperature { get; init; }
+    public float? Temperature { get; init; }
 
     /// <summary>
     /// Request JSON format response (if supported by model)
